Fixing 2D array allocation

+++ PCI_Bandwidth_Benchmark/malloc2D.c
@@ -1,13 +1,13 @@
 #include <stdlib.h>
 #include "malloc2D.h"
 
-float **malloc2D(int jmax, int imax)
+double **malloc2D(int jmax, int imax)
 {
    // first allocate a block of memory for the row pointers and the 2D array
-   float **x = (float **)malloc(jmax*sizeof(float *) + jmax*imax*sizeof(float));
+   double **x = (double **)malloc(jmax*sizeof(double *) + jmax*imax*sizeof(double));
 
    // Now assign the start of the block of memory for the 2D array after the row pointers
-   x[0] = (float *)x + jmax;
+   x[0] = (double *)x + jmax;
 
    // Last, assign the memory location to point to for each row pointer
    for (int j = 1; j < jmax; j++) {
