+++ PCI_Bandwidth_Benchmark/malloc2D.h
@@ -1,1 +1,1 @@
-float **malloc2D(int jmax, int imax);
+double **malloc2D(int jmax, int imax);
